+++ src/serialization.ml
@@ -281,8 +281,8 @@ let rec value_to_yojson (v : Ast.value) : Yojson.Safe.t =
           | Some tz -> `String tz
           | None -> `Null));
       ]
-  | VFactor _ | VSerializer _ ->
-      invalid_arg "value_to_yojson: VFactor/VSerializer is not supported for JSON serialization"
+  | VFactor _ | VSerializer _ | VBuildLog _ ->
+      invalid_arg "value_to_yojson: VFactor/VSerializer/VBuildLog is not supported for JSON serialization"
   | VUnquote _ | VUnquoteSplice _ | VDynamicArg _ | VQuo _ | VEnv _ ->
       invalid_arg "value_to_yojson: metaprogramming intermediate values are not serializable"
   | VLens l ->